Use rooms to filter messages
Since the app should allow users to receive messages in multiple rooms
at a time, there is not really a concept of "leaving" the room.  This
means that using join/leave alone isn't enough to filter messages from
the server as they arrive at the client.  On the plus side, it cuts
down on traffic for the case where there are many rooms and the user
is only in a small subset.  However, the filter at room.js:39 needs
to stay.